+++ tokens/usecase/tokens_usecase_test.go
@@ -212,7 +212,7 @@ func (s *TokensUseCaseTestSuite) TestGetPrices_Chain() {
 	// in place against a hardcoded expected value rather than comparing USDT and USDC prices only
 	// that are both computed by the system.
 	// Noe: if WBTC price changes by more than 15% and we update test mainnet state, this test is likely to fail.
-	expectedwBTCPrice := osmomath.NewBigDec(56000)
+	expectedwBTCPrice := osmomath.NewBigDec(68000)
 	wbtcErrorTolerance := osmomath.ErrTolerance{
 		// 15% tolerance
 		MultiplicativeTolerance: osmomath.MustNewDecFromStr("0.15"),